modify transpezoidal speed.
